implementação do sistema de centro de gravidade

+++ Raspberry/src/com/br/phdev/cmp/GravitySystem.java
@@ -167,7 +167,7 @@ public class GravitySystem  {
         }
 
         @Override
-        public void onServoTaskComplete(float currentPos) {
+        public void onServoTaskComplete(double currentPos) {
             notifyAll();
         }
